requirements.txt
flask
flask-cors
mysql-connector-python
psycopg2-binary
gunicorn
